feat: Add initial binary for kavachat API
This creates a new go binary build that will run the API required for
deployment and management of an OpenAI proxy to support our frontend
chatbot client.

A basic E2E test for the binary is added to be expanded to support
configuration, arguments, and smoke testing requests with graceful
shutdown.