feat: Require OPENAI_API_KEY for kavachat api
The kavachat-api now checks for the existance of the OPENAI_API_KEY and
writes an error to stderr with an exit status of 1 if no key is set in
the environment.

This is done in order to catch invalid configurations quicker in the
startup process and always ensure some key is set in order to interact
with the OpenAI API.

The actual check to ensure the API key is valid and able to call OpenAI
can be or should be done in a future healthcheck endpoint to ensure
correct configuration in production settings and ensure that outbound
network connections are possible before the service is served traffic.