feat: Add structured logging with stdout error log
This adds a structured logger using the built-in slog logger that logs
to stdout and gives structured output to the openai key required error.

We also adjust the openai fatal error to log both to stdout and stderr.
Structured logging to stdout is important for downstream processes to
record the error level log, but stderr is also important for debugging
and allowing an easy visual when watching stderr into why the process
did not start.