feat: Add basic server with e2e test capabilities
This adds a server that used the KAVACHAT_API_PORT variable to determine
what port to listen too.  If this variable is set to 0, then the server
will listen on any avaliable port.  If this variable is unset, the
server will listen on port 5555.

This also adds a healthcheck endpoint and the start of an e2e setup
that starts the server with KAVACHAT_API_PORT=0, gets the listening port
from the stdout, and polls the healthcheck until the server is ready.

If the server does not start within the process context deadline, the
process and running go-routines are cleaned up appropiatly.  In
addition, un-recoverable errors are caught and causes the server to
shutdown.

This test will form the basis for e2e testing openai proxy handling,
in addition allow for multiple e2e tests to run in parallel without
sharing state.