feat: Add basic working proxy for chat completions
This adds a fully E2E tested /chat/completions proxy for openai
streaming and non-streaming requests.  The test suite uses a mock server
that uses recorded OpenAI requests to provide real responses, which can
be re-generated and updated.

The main work done here is in creating a test suite that can be expanded
for testing logging, rate limits, and server-side events in detail.

In addition, the proxy is decoupled from the tests so it can be
refactored and the implemenation can be changed freely as we harden it
for production usage.

The python integration tests where modified to get them to pass. First
the request id assertion was removed since we won't be passing that back
tot he client, and the baseurl was modified to include the api version
which matches the expectations for the OPENAI_BASE_URL variable.