fix: Disable transfer encoding and use header Set
When using server-sent events, transfer encoding should not be used.
Chunked transfer encoding causes buffering issues and is automatically
set by the golang server when sending streaming responses.

We add a test to ensure that transfer-encoding is not set in the
response and this is done by setting the header value to 'identity'.

In addition, out of caution, we modify header values to use the Set
method over the Add method to catch errors when setting a header
multiple times which can be done when adding middleware, etc.