feat: Add Python OpenAI client integration tests (#2)
This adds two integration tests -- one for streaming and one
non-streaming requests.  This ensures that the proxy handles both
request-reponse and server side events correctly when using the Python
offical client.

The tests support two environment variables for configuration.

BYPASS_PROXY when set to 'true', will run the tests directly against
OpenAI with the OPENAI_API_KEY set in the environment.  This allows
us to ensure the tests are correct and isolate proxy failures.

PROXY_ENDPOINT which defaults to 'http://localhost:5555/openai', and
gives the full scheme://host:port/path to the proxy endpoint that
forwards requests to the open ai platform.