feat: Add test golden option for generating data
The KAVACHAT_TEST_GOLDEN env variable when set to 'true' will now
regenerate the openai saved test data for the mock server.

By default, the already stored responses are used and not regenerated,
but this is useful when adding new tests, changing them, etc.

In addition, fixed the duplicated env var logic of fetching the openai
key and base url so tests now do not error when run in an environment
without an API key or base url set.