feat: Add basic proxy implementation for openai chat completion (#9)
* feat: Add initial binary for kavachat API

This creates a new go binary build that will run the API required for
deployment and management of an OpenAI proxy to support our frontend
chatbot client.

A basic E2E test for the binary is added to be expanded to support
configuration, arguments, and smoke testing requests with graceful
shutdown.

* feat: Require OPENAI_API_KEY for kavachat api

The kavachat-api now checks for the existance of the OPENAI_API_KEY and
writes an error to stderr with an exit status of 1 if no key is set in
the environment.

This is done in order to catch invalid configurations quicker in the
startup process and always ensure some key is set in order to interact
with the OpenAI API.

The actual check to ensure the API key is valid and able to call OpenAI
can be or should be done in a future healthcheck endpoint to ensure
correct configuration in production settings and ensure that outbound
network connections are possible before the service is served traffic.

* feat: Add structured logging with stdout error log

This adds a structured logger using the built-in slog logger that logs
to stdout and gives structured output to the openai key required error.

We also adjust the openai fatal error to log both to stdout and stderr.
Structured logging to stdout is important for downstream processes to
record the error level log, but stderr is also important for debugging
and allowing an easy visual when watching stderr into why the process
did not start.

* refactor: Extract fatal logging to function

The fatal logging behavior of writing structured output to stdout and
stderr while exiting, should be encapsulated together and is an outside
concern of the higher-level main function.

* refactor: Use structured logging for startup

Convert the startup message to a structured info format

* feat: Add basic server with e2e test capabilities

This adds a server that used the KAVACHAT_API_PORT variable to determine
what port to listen too.  If this variable is set to 0, then the server
will listen on any avaliable port.  If this variable is unset, the
server will listen on port 5555.

This also adds a healthcheck endpoint and the start of an e2e setup
that starts the server with KAVACHAT_API_PORT=0, gets the listening port
from the stdout, and polls the healthcheck until the server is ready.

If the server does not start within the process context deadline, the
process and running go-routines are cleaned up appropiatly.  In
addition, un-recoverable errors are caught and causes the server to
shutdown.

This test will form the basis for e2e testing openai proxy handling,
in addition allow for multiple e2e tests to run in parallel without
sharing state.

* feat: Add basic working proxy for chat completions

This adds a fully E2E tested /chat/completions proxy for openai
streaming and non-streaming requests.  The test suite uses a mock server
that uses recorded OpenAI requests to provide real responses, which can
be re-generated and updated.

The main work done here is in creating a test suite that can be expanded
for testing logging, rate limits, and server-side events in detail.

In addition, the proxy is decoupled from the tests so it can be
refactored and the implemenation can be changed freely as we harden it
for production usage.

The python integration tests where modified to get them to pass. First
the request id assertion was removed since we won't be passing that back
tot he client, and the baseurl was modified to include the api version
which matches the expectations for the OPENAI_BASE_URL variable.

* fix: Disable transfer encoding and use header Set

When using server-sent events, transfer encoding should not be used.
Chunked transfer encoding causes buffering issues and is automatically
set by the golang server when sending streaming responses.

We add a test to ensure that transfer-encoding is not set in the
response and this is done by setting the header value to 'identity'.

In addition, out of caution, we modify header values to use the Set
method over the Add method to catch errors when setting a header
multiple times which can be done when adding middleware, etc.

* feat: Add test golden option for generating data

The KAVACHAT_TEST_GOLDEN env variable when set to 'true' will now
regenerate the openai saved test data for the mock server.

By default, the already stored responses are used and not regenerated,
but this is useful when adding new tests, changing them, etc.

In addition, fixed the duplicated env var logic of fetching the openai
key and base url so tests now do not error when run in an environment
without an API key or base url set.

* feat: Add very basic CORS support for local dev

Add a basic (and not fully to spec) CORS implementation to allow the
server to work locally and in CI.  The implementation needs to be
extended still to only allow preflight with an origin header set,
offer content-length for non streaming requests, add more standard
headers, add access control cachine (and possibly other caching), and
review other headers, spec, etc for full implementation.