refactor: Remove repetitive vitest imports
This adjusts the vitest configuration to use globals for BDD imports
like describe, it, beforeAll, etc instead of requiring these to be
imported for every test file.  In addition, tsconfig is updated to
ensure the types are included for these globals, and the integration
test also supports these globals.

All typescript test files are update to remove the imports that are now
global except for Mock, which is not a vitest global import.

Co-Authored-By: adamrturman <adamrturman@gmail.com>