feat: Add user driven system prompt for user types (#43)
Build a system prompt that is more direct and geared towards the
different user use-cases we'd like to support: New, New to Kava, and
Power User. Simply the prompt so it can be concise yet still robust
enough to help users. Provide some specifics in areas where the model
may halucinate such as the exchanges where KAVA is available.

Co-authored-by: Paul Downing <=>