feat: Clean up font spacing and sizing
This sets the base font to 10px == 1rem.  To use 16px font size, set to
1.6rem.

Font weights, line heights, and such are also set for readability and to
ensure we have good base reset set to build on top of.

In addition, set entire body background to match color we want and
prevent weird affects when content doesn't fill screen.