feat: Add storybook for component UI development
This adds storybook which is used to iterate on components in responsive
views, with different settings, etc. This enables faster development
when working with components in different states instead of having to
manually trigger the state in the full app.

This also allows components to be developed in isolation before being
included in the page.