feat: Add ChatView component for basis of demo
This adds a ChatView component that we can iterate on and integrate to
replace the current UI.  It's responsive with a 768px max width on large
screen, resistant to scrollbar shifts, and scales well to phone screen
heights and widths.

The tests are incomplete, adding rendering check for the messages next.