feat: Polish header CSS
This adds a responsive header that scales down well for mobile screens.
In addition, the header was changed from fixed positioning to sticky
position to have the same effect, but flow with the layout which solves
the conversation container being under the header and allows the header
to not interfere witht the scrollbar.