feat: New UX and Refactored State (#74)
* feat: Clean up font spacing and sizing

This sets the base font to 10px == 1rem.  To use 16px font size, set to
1.6rem.

Font weights, line heights, and such are also set for readability and to
ensure we have good base reset set to build on top of.

In addition, set entire body background to match color we want and
prevent weird affects when content doesn't fill screen.

* feat: Add storybook for component UI development

This adds storybook which is used to iterate on components in responsive
views, with different settings, etc. This enables faster development
when working with components in different states instead of having to
manually trigger the state in the full app.

This also allows components to be developed in isolation before being
included in the page.

* lint: Lint fix and prettier run on misc files

These files are from upstream, adding a separate commit for their fixes.

* fix: Remove incomplete import

This should not have been committed, removed the line.

* feat: Add ChatView component for basis of demo

This adds a ChatView component that we can iterate on and integrate to
replace the current UI.  It's responsive with a 768px max width on large
screen, resistant to scrollbar shifts, and scales well to phone screen
heights and widths.

The tests are incomplete, adding rendering check for the messages next.

* test: Add unit tests for chat view rendering

Covers no props, empty messages, and having other messages to cover
the logical cases in the component to ensure it's resilient against
refactors and prop changes.

* feat: Add nested Send btn within input with scoll

Add a nested send chat button within the in put window. Leave chat btn
not wired in yet. Add scroll ref to allow the textarea to scroll upward
and keep the user on the same line as they type past new lines

* feat: Limit input height to 60vh and remove ref

This enables the user input to start scrolling once it reaches a
majority of the screen height.

In addition, since we are not passing a reference of the input to
another component, we can drop the ref from the textarea input.

* feat: Add Reset Chat btn with empty click handler

Add a Reset Chat button in a sticky header so users are able to reset
their chats and start fresh. Currently the work to reset chats is not in
this branch so for now function will just do nothing

* fix: correct unit test with reset chat button

Reset Chat button now no longer shows when there are no messages, but
the test itself didn't get updated to follow logic. Fix that test.

* chore: Move SVG icons to simple react components

Move the reset chat and send chat icons to simple react components so
they are extracted away from the view code and for readability

* fix: Update ChatView stories to use empty message

ChatView component must have messages (mandatory prop), but the some of
the stories were missing this as props.

* feat: Polish header CSS

This adds a responsive header that scales down well for mobile screens.
In addition, the header was changed from fixed positioning to sticky
position to have the same effect, but flow with the layout which solves
the conversation container being under the header and allows the header
to not interfere witht the scrollbar.

* fix: Fix invalid properties for react svg

* refactor: Extract sanitize function for text

This extracts the marked.parse and DOMPurify logic into it's own
function outside of react components.  The async usage is used on
purpose to signify that this function should not be used in the react
render loop and should be used via useEffect or other async methods.

* fix: Fix type import in mockdata function

* fix: Prettier + fix typo in css

* wip: adding tests and re-reviewing code

* fix: reset package-lock.json

* feat: memoize expensive components (react profile based optimization) (#77)

* feat: memoize expensive components

* chore: add comment

* fix: playwright e2e test (#79)

---------

Co-authored-by: Paul Downing <=>
Co-authored-by: samcode206 <66322745+samcode206@users.noreply.github.com>