feat: wire launch btn to fill send event to iframe
Set a click handler on the launch button to postMessage to the parent
from the iframe. Use that data in the hard.fun demo to set the form
fields with avatar, name, symbol, and description

Setup a handler to clear the DOM fields when the demo html page is
refreshed.

Fix styling of Close Chat button to look like hard.fun page