Pd/hard fun demo (#84)
* feat: add launch btn and update action btn styling

Add a "Launch" button that is not yet wired into the hard.fun demo page.

Update styling for Send chat and reset chat buttons to match hard.fun

* fix: Fix center bug and style w/ send/cancel btns

Cancel and send chat buttons were not centered (a few pixels off). Fixed
styling to match hard.fun colors

* feat: build hard.fun demo with Hydra AI embed chat

Add view page to mimic hard.fun and show product what our chat would
look like once integrated into hard.fun website

* feat: Move chat demo files to their own sub dir

We want demo stuff out of the public folder so it's not really messing
with our pipeline and we can keep styling and other items which are
unique to the demo separate from the chat code base

* fix: remove no longer used div on hard.fun demo

* feat: wire launch btn to fill send event to iframe

Set a click handler on the launch button to postMessage to the parent
from the iframe. Use that data in the hard.fun demo to set the form
fields with avatar, name, symbol, and description

Setup a handler to clear the DOM fields when the demo html page is
refreshed.

Fix styling of Close Chat button to look like hard.fun page

* fix: Clean up tsx svg key warnings

stroke-width, stroke-linejoin, and stroke-linecap need to be Camel Case
instead of Kebab Case in jsx/tsx files or they'll throw a warning even
though they render correctly

* feat: Soften btn and icon colors to blend better

Add an opactiy to the orange rgb colors on icons so fit better with the
overall theme and align better with the main chat icon on hard.fun

* feat: Soften the launch button orange coloring

Soften the Launch button coloring to align with the rest of the them and
match the diamond logo as well as other buttons

* feat: Soften send chat and reset button colors

Soften the orange color on Send and Rest chat buttons to align with
other buttons, chatIcon, and hard.fun theming

* fix: add even padding to top of stickyHeader

Add additional padding to the top of the sticky header to both allow for
consistency of the padding in that element and also to allow a little
padding above the reset chat icon in the iframe

* feat: Swap out chatIcon for hard.fun diamond icon

Instead of using our chatIcon on the intro page for the chat, we want to
use the diamond icon to align with hard.fun branding. Add some styling
with the icon to blend it into the page and alter the color to hard.fun
orange with some opacity to soften it up

* feat: Replace old chatIcon with new chatIcon

Replace the old chatIcon svg with the new hardDotFunDiamond svg and
adjust the spacing where it is used in chats to make sure content still
aligns nicely

* fix: Fix chatIcon styling on main chat page

Since I decided to build the circular aspect directly into the svg, the
css code that was building the circule around the chatIcon was no longer
needed

* fix: adjust hardDotFunDiamond height/width

Update hardDotFunDiamond icon height and width so it is a perfect circle
This is needed for correct sizing with borders, etc. Adjust background
color of buttons to align with the new image background color

* feat: Add svgs to simplify html code in demo

These images were being used as raw svg code in the demo markup making
it messy. hardDotFunDiamond.svg is an exacty clone of the same named
file in assets, but is here so it doesn't get processed by vite and can
be used on the standalone page.

* fix: clean up code and simplify

Moved svg images out of the code and imported them instead. Updated js
script import to help prevent build issues

* fix: Update the svgs and open chat btn styles

Update the svgs in the javascript handlers to use the most recent svgs
for hardDotFunDiamond and defaultAvatar. Update button styling on close
chat button to align with other button styles

* fix: Update demo styling to fix misc UX bugs

- update background position in body to make sure page background
  extends all the way to the edges of the page
- hide box shadow for now on iframe for now but keep in code to bring
  back into the next PR
- adjust open_chat nesting to use img instead of svg directly since the
  html has changed to use an img tag
- postion iframe closer to the edge of the window to prevent it from
  covering up parts of the from
- add some styling to the iframe window to make it pop a bit more and
  fit better into the page
- fix typo in active textarea styles

---------

Co-authored-by: Paul Downing <=>