feat: redesign TokenCard comp for user clarity
Redeisgn the TokenCard component to help draw out and separate the diff
parts of info for the user. Adjust the error styling to fix with the
look and feel of the card component

In addition to not showing the launch button when loading, we also don't
want to show it when there is an image error or when the user is not in
the iframe as the launch button is tied to pushing the data up out and
out of the iframe

Clean up the css to use variabels and help simplify the markup.