feat: convert messages with masks back to original messages
Create a unit-tested function that takes in a string and a map of masks to addresses. The function replaces the occurrences of the mask in the message with its original value.

`Send 100 KAVA to address_1` will become `Send 100 KAVA to 0xd8e30f7bcb5211e591bbc463cdab0144e82dffe5` (provided that masksToValues in local storage has those values)

It also handles the edge case where masks are present in the input, but not in local storage (although I'm not sure there's a way to do that via our UI)