fix: delete from UI, then from local storage
This is similar to the change recently made for title editing. If a user clicks to delete a chat, we don't want to wait for a synchronous action to complete before updating the UI.

We will track conversations in local state and remove it from there first and then schedule the update to local storage.