feat: add qwen3-30b-a3b as available models (#542)
* feat: add qwen3-30b-a3b as available models

model is listed as "experimental". it is included in a way that makes it
available whether or not the VITE_FEAT_QWEN flag is set.

* fix: adjust styling of model selector dropdown

the menu item container now grows to properly contain the different
model descriptions

* fix: support parsing reasoning_content in delta

models can be served with a reasoning parser. our Qwen model uses one
which results no <think> tags in the main content. thinking content is
in the "reasoning_content" field of streamed chat completion deltas.